+++ Modules/Intent.Modules.Application.FluentValidation.Dtos/release-notes.md
@@ -1,3 +1,7 @@
+### Version 3.11.2
+
+- Improvement: Ability to set a a timeout on `Regular Expression` validations
+
 ### Version 3.11.1
 
 - Improvement: Abililty to set a custom message on `Regular Expression` and `Must` validations