+++ Modules/Intent.Modules.Application.FluentValidation/release-notes.md
@@ -1,6 +1,10 @@
+### Version 3.10.2
+
+- Improvement: Ability to set a a timeout on `Regular Expression` validations
+
 ### Version 3.10.1
 
-- Improvement: Abililty to set a custom message on `Regular Expression` and `Must` validations
+- Improvement: Ability to set a custom message on `Regular Expression` and `Must` validations
 
 ### Version 3.10.0